(possible) fix for macOS config not being sourced
as well as making the effort to keep main.rs tiny with the necessary
functions in separate modules.
