+++ src/error.rs
@@ -6,7 +6,7 @@ pub enum Error {
     #[error("Failed due to IOError {0}")]
     IOError(#[from] io::Error),
 
-    #[error("Failed to parse toml file {0}")]
+    #[error("Failed to parse TOML file {0}")]
     ParsingError(#[from] toml::de::Error),
 }
 
@@ -14,9 +14,9 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 pub fn print_errors(err: Error) {
     match err {
-        Error::ParsingError(err) => match err.line_col() {
-            Some((line, col)) => {
-                //  Indexes are 0-based, let's increment them to make it intuitive
+        Error::ParsingError(err) => match err.line_col() { Some((line, col)) => {
+                //  Indexes are 0-based, we increment
+                //  them to make them less confusing
                 println!(
                     "\x1b[31mError\x1b[0m: At line {} column {}\nCaused by: {}",
                     line + 1,
