+++ src/main.rs
@@ -1,118 +1,26 @@
 #![warn(clippy::all)]
+mod ascii;
 mod bars;
+mod buffer;
 mod cli;
 mod config;
+mod data;
+mod doctor;
 mod error;
 mod extra;
 mod format;
 pub mod theme;
+pub mod widgets;
 
 use cli::Opt;
-use std::io;
+use error::Result;
 use structopt::StructOpt;
+use tui::backend::Backend;
+use tui::buffer::Buffer;
+use tui::layout::Rect;
 
 #[macro_use]
 extern crate lazy_static;
-
-mod ascii;
-mod data;
-mod doctor;
-pub mod widgets;
-
-use crate::theme::Theme;
-use crate::widgets::readout::ReadoutList;
-use atty::Stream;
-use data::Readout;
-use error::Result;
-use std::io::Stdout;
-use tui::backend::{Backend, CrosstermBackend};
-use tui::buffer::{Buffer, Cell};
-use tui::layout::{Margin, Rect};
-use tui::text::Text;
-use tui::widgets::{Block, Borders, Paragraph, Widget};
-use unicode_width::UnicodeWidthStr;
-
-fn create_backend() -> CrosstermBackend<Stdout> {
-    CrosstermBackend::new(io::stdout())
-}
-
-fn find_widest_cell(buf: &Buffer, last_y: u16) -> u16 {
-    let area = &buf.area;
-    let mut widest: u16 = 0;
-    let empty_cell = Cell::default();
-
-    for y in 0..last_y {
-        for x in (0..area.width).rev() {
-            let current_cell = buf.get(x, y);
-            if current_cell.ne(&empty_cell) && x > widest {
-                widest = x;
-                break;
-            }
-        }
-    }
-
-    widest + 1
-}
-
-fn find_last_buffer_cell_index(buf: &Buffer) -> Option<(u16, u16)> {
-    let empty_cell = Cell::default();
-
-    if let Some((idx, _)) = buf
-        .content
-        .iter()
-        .enumerate()
-        .filter(|p| !(*(p.1)).eq(&empty_cell))
-        .last()
-    {
-        return Some(buf.pos_of(idx));
-    }
-
-    None
-}
-
-fn draw_ascii(ascii: Text<'static>, tmp_buffer: &mut Buffer) -> Rect {
-    let ascii_rect = Rect {
-        x: 1,
-        y: 1,
-        width: ascii.width() as u16,
-        height: ascii.height() as u16,
-    };
-
-    Paragraph::new(ascii).render(ascii_rect, tmp_buffer);
-    ascii_rect
-}
-
-fn draw_readout_data(data: Vec<Readout>, theme: Theme, buf: &mut Buffer, area: Rect) {
-    let mut list = ReadoutList::new(data, &theme);
-
-    if theme.get_block().is_visible() {
-        list = list
-            .block_inner_margin(Margin {
-                horizontal: theme.get_block().get_horizontal_margin(),
-                vertical: theme.get_block().get_vertical_margin(),
-            })
-            .block(
-                Block::default()
-                    .border_type(theme.get_block().get_border_type())
-                    .title(theme.get_block().get_title())
-                    .borders(Borders::ALL),
-            );
-    }
-
-    list.render(area, buf);
-}
-
-fn get_version() -> Result<()> {
-    if let Some(git_sha) = option_env!("VERGEN_GIT_SHA_SHORT") {
-        println!("macchina     {} ({})", env!("CARGO_PKG_VERSION"), git_sha);
-    } else {
-        println!("macchina     {}", env!("CARGO_PKG_VERSION"));
-    }
-
-    println!("libmacchina  {}", libmacchina::version());
-
-    Ok(())
-}
 
 fn main() -> Result<()> {
     let arg_opt = Opt::from_args();
@@ -127,10 +35,10 @@ fn main() -> Result<()> {
     }
 
     if opt.list_themes {
-        return Theme::list_themes(&opt);
+        return theme::list_themes(&opt);
     }
 
-    let theme = Theme::create_theme(&opt);
+    let theme = theme::create_theme(&opt);
     let should_display = data::should_display(&opt);
     let readout_data = data::get_all_readouts(&opt, &theme, should_display);
 
@@ -142,7 +50,7 @@ fn main() -> Result<()> {
     const MAX_ASCII_HEIGHT: usize = 50;
     const MINIMUM_READOUTS_TO_PREFER_SMALL_ASCII: usize = 8;
 
-    let mut backend = create_backend();
+    let mut backend = buffer::create_backend();
     let mut tmp_buffer = Buffer::empty(Rect::new(0, 0, 500, 50));
     let mut ascii_area = Rect::new(0, 1, 0, tmp_buffer.area.height - 1);
     let prefers_small_ascii =
@@ -160,24 +68,24 @@ fn main() -> Result<()> {
             }
 
             if ascii_art.width() != 0 && ascii_art.height() < MAX_ASCII_HEIGHT {
-                ascii_area = draw_ascii(ascii_art.to_owned(), &mut tmp_buffer);
+                ascii_area = buffer::draw_ascii(ascii_art.to_owned(), &mut tmp_buffer);
             }
         } else if prefers_small_ascii {
             // prefer smaller ascii in this case
             if let Some(ascii) = ascii::select_ascii(ascii::AsciiSize::Small) {
-                ascii_area = draw_ascii(ascii.to_owned(), &mut tmp_buffer);
+                ascii_area = buffer::draw_ascii(ascii.to_owned(), &mut tmp_buffer);
             }
         } else {
             // prefer bigger ascii otherwise
             if let Some(ascii) = ascii::select_ascii(ascii::AsciiSize::Big) {
-                ascii_area = draw_ascii(ascii.to_owned(), &mut tmp_buffer);
+                ascii_area = buffer::draw_ascii(ascii.to_owned(), &mut tmp_buffer);
             }
         }
     }
 
     let tmp_buffer_area = tmp_buffer.area;
 
-    draw_readout_data(
+    buffer::draw_readout_data(
         readout_data,
         theme,
         &mut tmp_buffer,
@@ -189,7 +97,7 @@ fn main() -> Result<()> {
         ),
     );
 
-    write_buffer_to_console(&mut backend, &mut tmp_buffer)?;
+    buffer::write_buffer_to_console(&mut backend, &mut tmp_buffer)?;
 
     backend.flush()?;
     print!("\n\n");
@@ -197,50 +105,14 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn write_buffer_to_console(
-    backend: &mut CrosstermBackend<Stdout>,
-    tmp_buffer: &mut Buffer,
-) -> Result<()> {
-    let term_size = backend.size().unwrap_or_default();
-
-    let (_, last_y) = find_last_buffer_cell_index(tmp_buffer)
-        .expect("An error occurred while writing to the terminal buffer.");
-
-    let last_x = find_widest_cell(tmp_buffer, last_y);
-
-    print!("{}", "\n".repeat(last_y as usize + 1));
-
-    let mut cursor_y: u16 = 0;
-    if atty::is(Stream::Stdout) {
-        cursor_y = backend.get_cursor().unwrap_or((0, 0)).1;
+fn get_version() -> Result<()> {
+    if let Some(git_sha) = option_env!("VERGEN_GIT_SHA_SHORT") {
+        println!("macchina     {} ({})", env!("CARGO_PKG_VERSION"), git_sha);
+    } else {
+        println!("macchina     {}", env!("CARGO_PKG_VERSION"));
     }
 
-    // we need a checked subtraction here, because (cursor_y - last_y - 1) might underflow if the
-    // cursor_y is smaller than (last_y - 1).
-    let starting_pos = cursor_y.saturating_sub(last_y).saturating_sub(1);
-    let mut skip_n = 0;
+    println!("libmacchina  {}", libmacchina::version());
 
-    let iter = tmp_buffer
-        .content
-        .iter()
-        .enumerate()
-        .filter(|(_previous, cell)| {
-            let curr_width = cell.symbol.width();
-            if curr_width == 0 {
-                return false;
-            }
-
-            let old_skip = skip_n;
-            skip_n = curr_width.saturating_sub(1);
-            old_skip == 0
-        })
-        .map(|(idx, cell)| {
-            let (x, y) = tmp_buffer.pos_of(idx);
-            (x, y, cell)
-        })
-        .filter(|(x, y, _)| *x < last_x && *x < term_size.width && *y <= last_y)
-        .map(|(x, y, cell)| (x, y + starting_pos, cell));
-
-    backend.draw(iter)?;
     Ok(())
 }
